+++ cypress/locators/anchor-navigation/locators.js
@@ -1,0 +1,3 @@
+// component preview locators
+export const ANCHOR_NAVIGATION = '[data-component="anchor-navigation"]';
+export const ANCHOR_NAVIGATION_STICKY_NAVIGATION = '[data-element="anchor-sticky-navigation"]';