+++ cypress/support/step-definitions/common-steps.js
@@ -1,5 +1,6 @@
 import {
-  visitComponentUrl, setSlidebar, pressESCKey, pressTABKey, asyncWaitForKnobs, visitFlatTableComponentNoiFrame,
+  visitComponentUrl, setSlidebar, pressESCKey, pressTABKey, asyncWaitForKnobs,
+  visitFlatTableComponentNoiFrame,
 } from '../helper';
 import {
   commonButtonPreview, labelPreview, helpIcon, helpIconByPosition, inputWidthSlider,
@@ -8,6 +9,7 @@ import {
   icon, inputWidthPreview, label, eventInAction, getDataElementByNameAndValue, storyRoot,
   precisionSlider, storyRootNoIframe, tooltipPreviewNoIframe, getDataElementByValueNoIframe,
   knobsNameTab, fieldHelpPreviewByPosition, labelByPosition, dlsRoot,
+  commonButtonPreviewNoIFrameRoot,
 } from '../../locators';
 import { dialogTitle, dialogSubtitle } from '../../locators/dialog';
 import { DEBUG_FLAG } from '..';
@@ -38,6 +40,10 @@ Given('I open {string} component page basic', (component) => {
 
 Given('I open {string} component page basic in iframe', (component) => {
   visitComponentUrl(component, 'basic', true);
+});
+
+Given('I open in full screen Test {string} component page in noIframe', (component) => {
+  visitComponentUrl(component, 'in_full_screen_dialog', true, 'test-');
 });
 
 Given('I open {string} component page buttonToogleGroup validation in iframe', (component) => {
@@ -222,6 +228,10 @@ When('I open component preview', () => {
   commonButtonPreview().click();
 });
 
+When('I open component preview in noIFrame', () => {
+  commonButtonPreviewNoIFrameRoot().click();
+});
+
 Then('component title on preview is {string}', (title) => {
   dialogTitle().should('have.text', title);
 });
@@ -385,12 +395,9 @@ Then('closeIcon is not visible', () => {
   closeIconButton().should('not.exist');
 });
 
-Then('closeIcon has the border outline', () => {
-  closeIconButton().should('have.css', 'outline', 'rgb(255, 181, 0) solid 3px');
-});
-
-Then('closeIcon has border outline for classic story', () => {
-  closeIconButton().should('have.css', 'outline', 'rgba(0, 103, 244, 0.247) auto 5px');
+Then('closeIcon has the border outline color {string} and width {string}', (color, width) => {
+  closeIconButton().should('have.css', 'outline-color', color)
+    .and('have.css', 'outline-width', width);
 });
 
 Then('closeIcon is focused', () => {
@@ -521,20 +528,8 @@ When('I press ESC on focused element', () => {
   cy.focused().trigger('keydown', { keyCode: 27, which: 27 });
 });
 
-When('I press Tab on focused element', () => {
-  cy.focused().trigger('keydown', { keyCode: 9, which: 9 });
-});
-
-When('I press Home on focused element', () => {
-  cy.focused().trigger('keydown', { keyCode: 36, which: 36 });
-});
-
-When('I press uparrow on focused element', () => {
-  cy.focused().trigger('keydown', { keyCode: 38, which: 38 });
-});
-
-When('I press End on focused element', () => {
-  cy.focused().trigger('keydown', { keyCode: 35, which: 35 });
+When('I press {word} on focused element', (key) => {
+  cy.focused().trigger('keydown', { key });
 });
 
 When('I press ShiftTab on focused element', () => {