+++ script/docker/cypress.sh
@@ -7,13 +7,14 @@ HELP_TEXT="Options:\n\n
 --build\t\t\tRuns the build tests\n
 --accessibility\t\tRuns the accessibility tests\n
 --visual\t\tRuns the visual tests\n
---regression\t\t\tRuns the regression tests
---regression-common\t\t\tRuns the regression common tests
---regression-experimental\t\t\tRuns the regression experimental tests
---regression-deprecated\t\t\tRuns the regression deprecated tests
---regression-validation\t\t\tRuns the regression validation tests
---regression-themes\t\t\tRuns the regression themes tests
---allure-report\t\t\tGenerates allure-report for cypress tests
+--regression\t\t\tRuns all regression tests\n
+--common\t\t\tRuns the regression common tests\n
+--experimental\t\t\tRuns the regression experimental tests\n
+--deprecated\t\t\tRuns the regression deprecated tests\n
+--validation\t\t\tRuns the regression validation tests\n
+--themes\t\t\tRuns the regression themes tests\n
+--test\t\t\t\tRuns the regression test directory tests\n
+--allure-report\t\tGenerates allure-report for cypress tests\n
 "
 
 RUN_BUILD=false
@@ -25,6 +26,7 @@ RUN_REGRESSION_EXPERIMENTAL=false
 RUN_REGRESSION_DEPRECATED=false
 RUN_REGRESSION_VALIDATION=false
 RUN_REGRESSION_THEMES=false
+RUN_REGRESSION_TEST=false
 RUN_ALLURE_REPORTS=false
 
 while test $# -gt 0; do
@@ -50,36 +52,42 @@ while test $# -gt 0; do
           shift
           ;;
         # run regression tests
+        # run all regression tests
         -r|--regression)
           RUN_REGRESSION=true
           shift
           ;;
         # run regression common tests
-        -rc|--regression-common)
+        -c|--common)
           RUN_REGRESSION_COMMON=true
           shift
           ;;
-       # run regression experimental tests
-       -re|--regression-experimental)
+        # run regression experimental tests
+        -e|--experimental)
           RUN_REGRESSION_EXPERIMENTAL=true
           shift
           ;;
-       # run regression deprecated tests
-        -rd|--regression-deprecated)
+        # run regression deprecated tests
+        -d|--deprecated)
           RUN_REGRESSION_DEPRECATED=true
           shift
           ;;
-       # run regression validation tests
-        -rv|--regression-validation)
+        # run regression validation tests
+        -v|--validation)
           RUN_REGRESSION_VALIDATION=true
           shift
           ;;
-       # run regression themes tests
-        -rt|--regression-themes)
+        # run regression themes tests
+        -t|--themes)
           RUN_REGRESSION_THEMES=true
           shift
           ;;
-       # run script to generate allure-report for cypress tests
+        # run regression > test directory > tests
+        -te|--test)
+          RUN_REGRESSION_TEST=true
+          shift
+          ;;
+        # run script to generate allure-report for cypress tests
         -ar|--allure-report)
           RUN_ALLURE_REPORTS=true
           shift
@@ -106,8 +114,8 @@ if $RUN_VISUAL; then
 fi
 
 if $RUN_REGRESSION; then
-  echo 'Running regression tests'
-  docker-compose exec -T cypress bash -c "wait-on http://storybook:9001 && npm run test-cypress-regression"
+  echo 'Running all regression tests'
+  docker-compose exec -T cypress bash -c "wait-on http://storybook:9001 && npm run test-cypress-regression; npm run generate-cypress-allure-report"
 fi
 
 if $RUN_REGRESSION_COMMON; then
@@ -133,6 +141,11 @@ fi
 if $RUN_REGRESSION_THEMES; then
   echo 'Running regression themes tests'
   docker-compose exec -T cypress bash -c "wait-on http://storybook:9001 && npm run cypress-group-regression-themes"
+fi
+
+if $RUN_REGRESSION_TEST; then
+  echo 'Running regression > test directory > tests'
+  docker-compose exec -T cypress bash -c "wait-on http://storybook:9001 && npm run cypress-group-regression-test"
 fi
 
 if $RUN_ALLURE_REPORTS; then